Move RunnerBase to own file to avoid import cycle
