+++ gdq/runners/__init__.py
@@ -1,28 +1,9 @@
 import argparse
 import sys
-from abc import ABC
-from abc import abstractmethod
-from datetime import datetime
 from typing import Dict
-from typing import Optional
 
-from gdq.events import MarathonBase
-from gdq.runners import bus
-from gdq.runners import gdq
-from gdq.runners import horaro
-
-
-class RunnerBase(ABC):
-    @abstractmethod
-    def get_marathon(self, event_config: dict, args: argparse.Namespace) -> Optional[MarathonBase]:
-        pass
-
-    @abstractmethod
-    def get_start(self, event_config: dict) -> datetime:
-        pass
-
-    def get_options(self, parser: argparse.ArgumentParser) -> argparse.Namespace:
-        return parser.parse_args()
+from gdq.runners import bus, gdq, horaro
+from gdq.runners.base import RunnerBase
 
 
 def get_base_parser() -> argparse.ArgumentParser:
@@ -34,6 +15,10 @@ def get_base_parser() -> argparse.ArgumentParser:
     parser.add_argument(
         "--oneshot", action="store_true",
         help="Run only once and then exit",
+    )
+    parser.add_argument(
+        "--list", action="store_true",
+        help="List all known events instead of tracking one",
     )
     parser.add_argument(
         "stream_name", nargs="?", type=str, default="gdq",
